requirements.txt
torch
transformers
peft
numpy
Pillow
argparse
einops
timm
xformers
sentencepiece